First draft - framework of most methods+Javadocs
This commit is actually a couple days worth of work, I just didn't
realize I could create a private repository for free as a student until
just now.

Most methods have the framework of what I'd like to do with them and the
comments.  I've begun to really dig into the logic of the dotProduct
method and I know for a fact that the current version doesn't work
properly, so that will soon change.