Added FileHandler, Main Class
I added the FileHandler class, which I will be using to write the
results of the algorithm to a text file.  I also moved the main method
out of my SRGsolver class into a Main class.

Additionally, within my SRGsolver class I tweaked the dotProduct method
to work correctly.

Finally, I added in a text file of my pre-programming plan.  A lot of
the logic work for the dotProduct method was done in my physical
notebook, but this text file gives a general view of my planning.