Updated file writing, other changes
I made some pretty major changes to the file writing classes so that
they print formatted tables.  This wasn't as easy as it sounds, as I had
to scale for different number of digits, which I didn't realize until I
thought I was done with it.  I also created a folder for text documents
of graphs that are done.

There are some other small miscellaneous changes that I'm sure I'm
forgetting, but the above are the drastic changes for this commit.