Updates to file writing, random row algorithm
In this commit I made some final changes with my file writing.
Additionally, I made a fairly significant change in the way random rows
are generated.  Using previous rows, we can find out the first (x)
values of the next random row.  I factored this concept into the random
row generating algorithm, which means that it will be more accurate in
it's row selection.  This will (likely) greatly improve the speed of
graph generation.

The updates to the way the files are written means I deleted my previous
text files of completed graphs.  These will be re-uploaded in the next
version in the slightly nicer format.

One final thing I added in this commit was an Excel spreadsheet
containing known graphs, and which of those known graphs I have
successfully generated with my algorithm.