Added complete graphs, small touch ups
It's been a while since I've made any changes to the actual algorithm;
I've spent the last couple months working on the website for the project
in between a busy work schedule.

That being said, in this commit I'll be added a lot more text files
containing the adjacency matrices and other information for some graphs.
Also, there have been some small touch ups made to the code in a few of
the files.